Refactor #216: Unify `add` and `import` into a single `add` command
`add` now accepts both a single EPUB file and a directory of EPUBs, and
dispatches internally on `path.is_file()` vs `path.is_dir()`. The mental
model "I want to add a book" no longer bifurcates by input shape — modern
CLIs (`uv add`, `npm install`, `gh repo clone`) dispatch on path type from
one verb; same pattern here.

- File path: copy into library_root, run match pipeline (default
  `--match`), catalog.
- Directory path: recursive scan + ingest, with `--no-match` default for
  large batch flows. `--convert`, `--force-duplicates`, `-o/--output-dir`
  carry over from the old `import` surface.
- `import` becomes a deprecated alias via the shared
  `deprecated_command_alias` helper (#221); it prints a one-line stderr
  warning and forwards all args/options to `add` unchanged. Removed in a
  future release.

README and docs updated. Tests cover both dispatch paths, missing-path
errors, the directory-only flags, and the deprecated `import` alias for
both file and directory inputs.

Closes #216.
