+++ docs/config.example.toml
@@ -7,7 +7,7 @@ library_root = "/Users/you/.library"
 # ---------------------------------------------------------------------------
 # PDF → EPUB semantic conversion
 #
-# `bookery add` / `bookery import` pipe PDFs through a single LLM call that
+# `bookery add` pipes PDFs through a single LLM call that
 # reassembles articles/chapters into a clean EPUB. Any OpenAI-compatible
 # endpoint works — LM Studio, OpenAI, Moonshot/Kimi, Anthropic-via-proxy, etc.
 # ---------------------------------------------------------------------------
